+++ tests/insert.rs
@@ -11,7 +11,7 @@ use tracing::debug;
 pub struct UserEntity {
     #[PrimaryKey]
     #[AutoIncrement]
-    id: Option<i32>,
+    id: Option<i64>,
     age: Option<i32>,
     name: String,
 }
